requirements.txt
Flask==2.3.3
bcrypt==4.0.1
python-dotenv==1.0.0
mysql-connector-python==8.1.0
Werkzeug==2.3.7
Flask-WTF==1.2.1
reportlab==4.4.4
Pillow==11.1.0